Feature: Webpack integration
