+++ index.html
@@ -8,10 +8,6 @@
 </head>
 <body></body>
 
-<script src="js/const.js"></script>
-<script src="js/rover.js"></script>
-<script src="js/grid.js"></script>
-<script src="js/mars.js"></script>
-<script src="js/app.js"></script>
+<script src="js/main.js"></script>
 
 </html>
