--- js/const.js
@@ -1,2 +1,0 @@
-const obsChar = 'O';
-const roverChar = 'R';
