+++ package.json
@@ -10,15 +10,22 @@
         "type": "git",
         "url": "https://github.com/mauriciogzrs/mars-rover.git"
     },
+    "main": "./app.js",
     "scripts": {
-        "start": "node js/node/app.js",
-        "lint": "eslint ./",
-        "lint-fix": "eslint ./ --fix",
-        "test": "mocha --recursive ./tests/"
+        "start": "node ./app.js",
+        "lint": "eslint .",
+        "lint-fix": "eslint --fix .",
+        "test": "mocha --recursive ./tests/",
+        "build": "webpack"
     },
     "devDependencies": {
+        "@babel/core": "^7.9.0",
+        "@babel/preset-env": "^7.9.5",
+        "babel-loader": "^8.1.0",
         "eslint": "^6.8.0",
         "eslint-config-google": "^0.14.0",
-        "mocha": "^7.1.1"
+        "mocha": "^7.1.1",
+        "webpack": "^4.42.1",
+        "webpack-cli": "^3.3.11"
     }
 }
